move gadget tests
